refactor: 날짜 관련 유틸리티 함수 하나의 책임만 가질 수 있도록 코드 수정 및 디렉토리 구조 수정
formatDateWithDayOfWeek.ts -> formatDate, getDayOfWeek 두개의 함수로 분리
Date 관련 유틸리티 함수 utils/dateUtils 디렉토리에서 관리하도록 폴더 구조 수정
